pyproject.toml

[project]
name = "payment-reminders"
version = "0.1.0"
description = ""
authors = [
    {name = "Monica Canaza",email = "monihikaricanaza@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"


dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "supabase (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "heyoo (>=0.1.2,<0.2.0)",
    "ipdb (>=0.13.13,<0.14.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlacodegen (>=2.3.0.post1,<3.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
